Debian::Control::FromCPAN: find_debs_for_modules(): exit version loop.
When searching for versioned dependency, the method issues the warning
"$module package in APT ($d) does not satisfy $dep"
even when higher versions are available.

The reason is that the change in 51f6896 loops over all versions of a
package in AptPkg::Cache in descending order, and then may warn even if a
good enough version has been seen before.

Exit the loop as soon as we have a high enough version.

Caveat: This version can be from any suite, since AptPkg::Cache and
AptPkg::Config don't seem to offer information about suites.