DhMakePerl::Utils: when checking for core modules, look at current perl.
Various helpers which check if a module is in Perl core or starting from
which version only looked at the first Perl version. This ignored cases
where a package was removed from core at some point.

Change those methods to return if a module is not currently in Perl core.

Thanks: Peter Valdemar Mørch for the bug report.
Closes: #788198