Adapt to missing /^FILE\s+LOCATION/ line in contents files
For some reason unknown to me, suddenly somewhere between Jan 14, 2016
and Jan 20, 2017, contents files in the official debian mirrors no
longer start with pre-able ending in a line matching /^FILE\s+LOCATION/.

lib/Debian/AptContents.pm had a $capturing flag in sub read_cache that
essentially skipped that first pre-amble.

The result of this was that no lines in contents files were picked up,
and we started seeing messages like:

    Parsing /var/cache/apt/apt-file/ftp.be.debian.org_debian_dists_jessie_main_Contents-amd64.gz ...
    Parsing /var/cache/apt/apt-file/ftp.be.debian.org_debian_dists_jessie_main_Contents-amd64.gz ...
    No APT contents can be loaded.
    Please install 'apt-file' package (at least version 2.5.0) and
    run 'apt-file update' as root.
    Dependencies not updated.

But there is also a "next unless expected line" functionality in the
loop over file lines that achieved the same thing. So I've removed the
$capturing flag.

This is not a problem "master" of dh-make-perl, because the entire
logic with $capturing flag and all was replaced with apt-helper in:

    commit 885b31c44b4a61d6f6ca44d3335c20506ab41ee9
    Author: gregor herrmann <gregoa@debian.org>
    Date:   Wed Mar 23 23:31:28 2016 +0100

        /usr/lib/apt/apt-helper cat-file can take multiple files as arguments

        this saves the loop over the Contents files

        thanks, nthykier

        Gbp-Dch: Ignore

Alas, apt-helper in debian stable / jessie isn't new enough for that
and I don't have guts to update the apt package on my jessie boxes :-)